Make BillingAddress optional
Make BillingAddress a pointer, so it can be omitted if no update is needed.

Also changes some uint32 to int for convenience.

+++ bigcommerce/orders_test.go
@@ -298,7 +298,7 @@ func TestOrderService_Edit(t *testing.T) {
 		Endpoint: "https://example.com",
 		UserName: "go-bigcommerce",
 		Password: "12345"})
-	statusID := uint32(1)
+	statusID := 1
 	params := &OrderEditParams{
 		StatusID: &statusID,
 	}
